+++ test/run_all_versions.sh
@@ -173,6 +173,7 @@ SUITES=(
 	objstore_addressing
 	objstore_allowlist
 	objstore_credentials
+	objstore_crlf
 	objstore_http_read
 	objstore_listing
 	objstore_module